@@ -1,10 +1,13 @@
 package com.project.template.exception;
 
+import java.io.Serial;
+
 public class EntityNotFoundException extends RuntimeException {
 
     /**
      * serial ID.
      */
+    @Serial
     private static final long serialVersionUID = 1L;
 
     /**
@@ -13,7 +16,7 @@ public class EntityNotFoundException extends RuntimeException {
      * @param id       the ID of the requested entity
      */
     public EntityNotFoundException(final Long id) {
-        super("Couldnt find entity with id: " + id);
+        super("Couldn't find entity with id: " + id);
     }
 
 }
