@@ -1,5 +1,6 @@
 package com.project.template.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
 
@@ -12,18 +13,19 @@ public final class SampleService extends BaseService<Sample, SampleResource> {
     /**
      * @param repository the user repository
      */
+    @Autowired
     public SampleService(final JpaRepository<Sample, Long> repository) {
         super(repository);
     }
 
     @Override
     protected SampleResource convertModelToResource(final Sample model) {
-        return getModelMapper().map(model, SampleResource.class);
+        return getMapper().map(model, SampleResource.class);
     }
 
     @Override
     protected Sample convertResourceToModel(final SampleResource resource) {
-        return getModelMapper().map(resource, Sample.class);
+        return getMapper().map(resource, Sample.class);
     }
 
 }
